Move link-params store slice to a shared space
