+++ packages/js/src/shared-admin/store/index.js
@@ -1,0 +1,1 @@
+export * from "./link-params";
