+++ packages/js/src/shared-admin/store/link-params.js
@@ -1,6 +1,8 @@
 import { createSelector, createSlice } from "@reduxjs/toolkit";
-import { get } from "lodash";
 import { addQueryArgs } from "@wordpress/url";
+import { get } from "lodash";
+
+export const LINK_PARAMS_NAME = "linkParams";
 
 /**
  * @returns {Object} The initial state.
@@ -8,14 +10,14 @@ import { addQueryArgs } from "@wordpress/url";
 export const createInitialLinkParamsState = () => get( window, "wpseoScriptData.linkParams", {} );
 
 const slice = createSlice( {
-	name: "linkParams",
+	name: LINK_PARAMS_NAME,
 	initialState: createInitialLinkParamsState(),
 	reducers: {},
 } );
 
 const linkParamsSelectors = {
-	selectLinkParam: ( state, linkParam, defaultValue = {} ) => get( state, `linkParams.${ linkParam }`, defaultValue ),
-	selectLinkParams: state => get( state, "linkParams", {} ),
+	selectLinkParam: ( state, linkParam, defaultValue = {} ) => get( state, `${ LINK_PARAMS_NAME }.${ linkParam }`, defaultValue ),
+	selectLinkParams: state => get( state, LINK_PARAMS_NAME, {} ),
 };
 linkParamsSelectors.selectLink = createSelector(
 	[
@@ -29,4 +31,4 @@ export { linkParamsSelectors };
 
 export const linkParamsActions = slice.actions;
 
-export default slice.reducer;
+export const linkParamsReducer = slice.reducer;
